laird: mwlwifi: Fix assoc finalization bug
When trying to connect, the last step is setting the AP MAC into the
firmware. The routines doing the HOSTCMD_CMD_SET_NEW_STN do a second
set (that incidentally includes our own MAC) reusing the buffer. Because
the buffer was also used to return the return code, the command has the
high bit set. It seems the sdio firmware filters the command out,
resulting in a time-out, error and ultimately keeping the higher levels
of the code in an unconnected state.

This patch corrects it by fixing the command field back to the command.

Bug: 10604